ENH: add basic R package
This PR adds a barebones R package for modeling heterogeneous treatment
effects.

Included are a few simple modes, different basis, single model, and transformed
outcome.

A method for computing and plotting the uplift curve and plotting the
actual vs predicted treatment effect by quantile.

A basic formula interface for specifing models of the form:
`outcome ~ x1 + x2 + ... + xn | treatment`. This still needs to be
cleaned up and tested a bit.